Move the config logic out of the mod.rs file
